create seperate js files for utlitites and td template and td update functionality
