+++ lib/index.js
@@ -3,11 +3,14 @@ require('./styles');
 require('../css/main');
 require('../css/reset');
 
-let $ = require('jquery');
-let TD = require('./new-td');
+const $ = require('jquery');
+const TD = require('./new-td');
 const storage = require('./storage');
+const tdTemplate = require('./td-template');
+const tdUpdate = require('./td-updates');
+const utilities = require('./td-utilities');
 
-let tdCount = 100;
+// let tdCount = 100;
 let tdLoadComplete = false;
 
 function showTenToDos() {
@@ -20,7 +23,7 @@ function showTenToDos() {
 }
 
 $(document).ready(function(){
-  loadStorage(tdLoadComplete);
+  utilities.loadStorage(tdLoadComplete);
   $('#save-button').on('click', formatNewTD);
   showTenToDos();
   $('#title-input').focus();
@@ -31,96 +34,96 @@ $("#show-more-todos").on('click', function(){
     $(".td-box").slice(showing - 1, showing + 10).show();
 });
 
-function tdTemplate (tdCreator, obj) {
-  let tdPageText = tdCreator;
-  let deleteButton = createDeleteButton();
-  let upVote = createUpVoteButton(obj);
-  let downVote = createDownVoteButton(obj);
-  let completeTD = createCompleteTDButton(obj);
-  prependTDs(obj, tdPageText, deleteButton, upVote, downVote, completeTD);
-  deleteTDsListener(tdPageText, deleteButton, upVote, downVote, completeTD);
-}
+// function tdTemplate (tdCreator, obj) {
+//   let tdPageText = tdCreator;
+//   let deleteButton = createDeleteButton();
+//   let upVote = createUpVoteButton(obj);
+//   let downVote = createDownVoteButton(obj);
+//   let completeTD = createCompleteTDButton(obj);
+//   prependTDs(obj, tdPageText, deleteButton, upVote, downVote, completeTD);
+//   deleteTDsListener(tdPageText, deleteButton, upVote, downVote, completeTD);
+// }
 
 function formatNewTD() {
-  let end = 2;
-  $( "article" ).slice( 0, end ).show();
+  // let end = 2;
+  // $( "article" ).slice( 0, end ).show();
+  let tdCount = utilities.getTDCount();
   let newTD = new TD(tdCount, $('#title-input').val(), $('#task-input').val());
-  tdTemplate(createTD(newTD), newTD);
+  tdTemplate.tdCreator(tdTemplate.tdHTML(newTD), newTD);
   storage.setTD(tdCount, newTD);
-  clearFields();
+  utilities.clearFields();
+  utilities.incrementTDCount();
   resetCounters();
-  tdCount++;
-  storage.setTDCount(tdCount);
   showTenToDos();
   $('#title-input').focus();
 }
 
-function createTD(obj) {
-  if (obj.completed === true) {
-    return $("<article id=" + obj.id + " class='td-box completed-td'>"
-        + "<div class='flexer'>"
-          + "<h2 class='td-title' contenteditable='true'>" + obj.title + "</h2>"
-        + "</div>"
-        + "<p class='td-task' contenteditable='true'>" + obj.task + "</p>"
-        + "<div class='importance-container'>"
-        + "<h4 class='importance-rating'>" + obj.importance + "</h4></div></article>");
-    } else return $("<article id=" + obj.id + " class='td-box td-incomplete'>"
-      + "<div class='flexer'>"
-        + "<h2 class='td-title' contenteditable='true'>" + obj.title + "</h2>"
-      + "</div>"
-      + "<p class='td-task' contenteditable='true'>" + obj.task + "</p>"
-      + "<div class='importance-container'>"
-      + "<h4 class='importance-rating'>" + obj.importance + "</h4></div></article>"
-    );
-}
+// function createTD(obj) {
+//   if (obj.completed === true) {
+//     return $("<article id=" + obj.id + " class='td-box completed-td'>"
+//         + "<div class='flexer'>"
+//           + "<h2 class='td-title' contenteditable='true'>" + obj.title + "</h2>"
+//         + "</div>"
+//         + "<p class='td-task' contenteditable='true'>" + obj.task + "</p>"
+//         + "<div class='importance-container'>"
+//         + "<h4 class='importance-rating'>" + obj.importance + "</h4></div></article>");
+//     } else return $("<article id=" + obj.id + " class='td-box td-incomplete'>"
+//       + "<div class='flexer'>"
+//         + "<h2 class='td-title' contenteditable='true'>" + obj.title + "</h2>"
+//       + "</div>"
+//       + "<p class='td-task' contenteditable='true'>" + obj.task + "</p>"
+//       + "<div class='importance-container'>"
+//       + "<h4 class='importance-rating'>" + obj.importance + "</h4></div></article>"
+//     );
+// }
 
-function createDeleteButton() {
-  return $("<button aria-label='delete to do' tabindex='0' class='delete-button'></button>");
-}
+// function createDeleteButton() {
+//   return $("<button aria-label='delete to do' tabindex='0' class='delete-button'></button>");
+// }
 
-function createUpVoteButton(obj) {
-  if (obj.completed === true) {
-    return $("<button aria-label='up vote todo' tabindex='0' class='up-button' disabled></button>");
-    } else return $("<button aria-label='up vote todo' tabindex='0' class='up-button'></button>");
-}
-
-
-function createDownVoteButton(obj) {
-  if (obj.completed ===true) {
-    return $("<button aria-label='down vote todo' tabindex='0' class='down-button' disabled></button>");
-  } else return $("<button aria-label='down vote todo' tabindex='0' class='down-button'></button>");
-}
-
-function createCompleteTDButton(obj) {
-  let tdID = obj.id;
-  let completeStatus = obj.completed;
-  if (completeStatus === true) {
-    return $("<button aria-label='complete to do' tabindex='0' class='completed-button' disabled>Completed</button>");
-  } else return $("<button aria-label='complete to do' tabindex='0' class='completed-button'>Completed</button>");
-}
-
-function prependTDs(obj, td, deleteButton, upVote, downVote, completeTD) {
-  if (obj.completed === true) {
-  $('.td-container').prepend(td);
-  td.find('.flexer').prepend(completeTD, deleteButton);
-  td.find('.importance-container').prepend(upVote, downVote);
-} else {
-  $('.not-complete-tds').prepend(td);
-  td.find('.flexer').prepend(completeTD, deleteButton);
-  td.find('.importance-container').prepend(upVote, downVote);
-  }
-}
-
-function deleteTDsListener(td, deleteButton, upVote, downVote, completeTD) {
-  deleteButton.on('click', function() {
-    td.remove();
-    deleteButton.off().remove();
-    upVote.off().remove();
-    downVote.off().remove();
-    completeTD.off().remove();
-    storage.removeTD(td.attr('id'));
-  });
-}
+// function createUpVoteButton(obj) {
+//   if (obj.completed === true) {
+//     return $("<button aria-label='up vote todo' tabindex='0' class='up-button' disabled></button>");
+//     } else return $("<button aria-label='up vote todo' tabindex='0' class='up-button'></button>");
+// }
+//
+//
+// function createDownVoteButton(obj) {
+//   if (obj.completed ===true) {
+//     return $("<button aria-label='down vote todo' tabindex='0' class='down-button' disabled></button>");
+//   } else return $("<button aria-label='down vote todo' tabindex='0' class='down-button'></button>");
+// }
+//
+// function createCompleteTDButton(obj) {
+//   let tdID = obj.id;
+//   let completeStatus = obj.completed;
+//   if (completeStatus === true) {
+//     return $("<button aria-label='complete to do' tabindex='0' class='completed-button' disabled>Completed</button>");
+//   } else return $("<button aria-label='complete to do' tabindex='0' class='completed-button'>Completed</button>");
+// }
+//
+// function prependTDs(obj, td, deleteButton, upVote, downVote, completeTD) {
+//   if (obj.completed === true) {
+//   $('.td-container').prepend(td);
+//   td.find('.flexer').prepend(completeTD, deleteButton);
+//   td.find('.importance-container').prepend(upVote, downVote);
+// } else {
+//   $('.not-complete-tds').prepend(td);
+//   td.find('.flexer').prepend(completeTD, deleteButton);
+//   td.find('.importance-container').prepend(upVote, downVote);
+//   }
+// }
+//
+// function deleteTDsListener(td, deleteButton, upVote, downVote, completeTD) {
+//   deleteButton.on('click', function() {
+//     td.remove();
+//     deleteButton.off().remove();
+//     upVote.off().remove();
+//     downVote.off().remove();
+//     completeTD.off().remove();
+//     storage.removeTD(td.attr('id'));
+//   });
+// }
 
 $('.td-container').on('click', '.completed-button', function() {
     let id = $(this).closest('.td-box').attr('id');
@@ -128,23 +131,23 @@ $('.td-container').on('click', '.completed-button', function() {
     $(currentTD).addClass('completed-td');
     $(currentTD).removeClass('.td-incomplete');
     $(this).attr('disabled', true);
-    updateTDStorage(id, 'completed', true);
-    checkButtons(storage.getTD(id));
+    tdUpdate.updateTDStorage(id, 'completed', true);
+    utilities.checkButtons(storage.getTD(id));
 });
 
-function updateTDStorage(id, attribute, newValue) {
-  let currentTD = storage.getTD(id);
-  if (attribute === 'importance') {
-    currentTD.quality = newValue;
-  } else if (attribute === 'title') {
-    currentTD.title = newValue;
-  } else if (attribute === 'task') {
-    currentTD.task = newValue;
-  } else if (attribute === 'completed') {
-    currentTD.completed = newValue;
-  }
-  storage.setTD(id, currentTD);
-}
+// function updateTDStorage(id, attribute, newValue) {
+//   let currentTD = storage.getTD(id);
+//   if (attribute === 'importance') {
+//     currentTD.quality = newValue;
+//   } else if (attribute === 'title') {
+//     currentTD.title = newValue;
+//   } else if (attribute === 'task') {
+//     currentTD.task = newValue;
+//   } else if (attribute === 'completed') {
+//     currentTD.completed = newValue;
+//   }
+//   storage.setTD(id, currentTD);
+// }
 
 $('.td-container').on('click', '.up-button, .down-button', function() {
   let parentSelector = $(this).closest(".td-box");
@@ -152,16 +155,17 @@ $('.td-container').on('click', '.up-button, .down-button', function() {
   let storedTD = storage.getTD(id);
   let currentImportance = storedTD.importance;
   if ($(this).is(".up-button")) {
-    storedTD.importance = upVoteChange(currentImportance);
+    storedTD.importance = tdUpdate.upVoteChange(currentImportance);
   } else if ($(this).is(".down-button")) {
-    storedTD.importance = downVoteChange(currentImportance);
+    storedTD.importance = tdUpdate.downVoteChange(currentImportance);
   }
-  changeImportanceText(parentSelector, storedTD.importance);
-  checkButtons(storedTD);
+  tdUpdate.changeImportanceText(parentSelector, storedTD.importance);
+  utilities.checkButtons(storedTD);
   storage.setTD(id, storedTD);
 });
 
 //need to figure out how to unfilter - either with a click on the sme button again OR with a "show all" - think we should also add some CSS to show the selected button highlighted/selected
+//we need to figure out how to ONLY show visible TDs when completed TDs are not shown...right now, both types of TDs are displaying when you hit the filter
 $('#filter-container').on('click', '#none-filter-button, #low-filter-button, #normal-filter-button, #high-filter-button, #critical-filter-button', function() {
   // let filterToggle;
   for (let i=0; i < localStorage.length; i++) {
@@ -169,95 +173,95 @@ $('#filter-container').on('click', '#none-filter-button, #low-filter-button, #no
     let storedTD = storage.getTD(localStorage.key(i));
     let storedTDElement = $(`#${storedTD.id}`);
     let storedTDImportance = storedTD.importance;
-    toggleShowHide(clickedButtonImportance, storedTDImportance, storedTDElement);
+    utilities.toggleShowHide(clickedButtonImportance, storedTDImportance, storedTDElement);
   }
 });
 
-function toggleShowHide(clickedButtonImportance, storedTDImportance, storedTDElement) {
-  if (clickedButtonImportance === storedTDImportance) {
-    storedTDElement.show();
-  }
-  else {
-    storedTDElement.hide();
-  }
-}
+// function toggleShowHide(clickedButtonImportance, storedTDImportance, storedTDElement) {
+//   if (clickedButtonImportance === storedTDImportance) {
+//     storedTDElement.show();
+//   }
+//   else {
+//     storedTDElement.hide();
+//   }
+// }
 
-function changeImportanceText(parentSelector, newImportance) {
-  parentSelector.find('.importance-rating').text(newImportance);
-}
+// function changeImportanceText(parentSelector, newImportance) {
+//   parentSelector.find('.importance-rating').text(newImportance);
+// }
+//
+// function upVoteChange(currentImportance) {
+//   switch (currentImportance) {
+//     case '5) None':
+//     return '4) Low';
+//     case '4) Low':
+//     return '3) Normal';
+//     case '3) Normal':
+//     return '2) High';
+//     case '2) High':
+//     return '1) Critical';
+//     default:
+//     return '1) Critical';
+//   }
+// }
+//
+// function downVoteChange(currentImportance) {
+//   switch (currentImportance) {
+//     case '1) Critical':
+//     return '2) High';
+//     case '2) High':
+//     return '3) Normal';
+//     case '3) Normal':
+//     return '4) Low';
+//     case '4) Low':
+//     return '5) None';
+//     default:
+//     return '5) None';
+//   }
+// }
 
-function upVoteChange(currentImportance) {
-  switch (currentImportance) {
-    case '5) None':
-    return '4) Low';
-    case '4) Low':
-    return '3) Normal';
-    case '3) Normal':
-    return '2) High';
-    case '2) High':
-    return '1) Critical';
-    default:
-    return '1) Critical';
-  }
-}
+// function resetTDCount() {
+//   let storedTDCount = storage.getTDCount();
+//   if (storedTDCount !== null) {
+//     tdCount = storedTDCount;
+//   }
+// }
 
-function downVoteChange(currentImportance) {
-  switch (currentImportance) {
-    case '1) Critical':
-    return '2) High';
-    case '2) High':
-    return '3) Normal';
-    case '3) Normal':
-    return '4) Low';
-    case '4) Low':
-    return '5) None';
-    default:
-    return '5) None';
-  }
-}
-
-function resetTDCount() {
-  let storedTDCount = storage.getTDCount();
-  if (storedTDCount !== null) {
-    tdCount = storedTDCount;
-  }
-}
-
-function loadStorage() {
-  for (let i = 0; i < localStorage.length; i++) {
-    if (localStorage.key(i) !== 'tdCountTracker') {
-      let storedTD = storage.getTD(localStorage.key(i));
-        tdTemplate(createTD(storedTD), storedTD);
-        checkButtons(storedTD);
-      }
-  }
-  resetTDCount();
-}
+// function loadStorage() {
+//   for (let i = 0; i < localStorage.length; i++) {
+//     if (localStorage.key(i) !== 'tdCountTracker') {
+//       let storedTD = storage.getTD(localStorage.key(i));
+//         tdTemplate.tdCreator(tdTemplate.tdHTML(storedTD), storedTD);
+//         utilities.checkButtons(storedTD);
+//       }
+//   }
+//   utilities.resetTDCount();
+// }
 
 $('#show-completed-todos').on('click', function() {
   tdLoadComplete = true;
   showTenToDos();
 });
 
-//we need to update the code here to make sure we are disabling the down/up buttons for completed tasks
-function checkButtons(obj) {
-  let id = obj.id;
-  let $completeButton = $(`#${id}`).children().children('.completed-button');
-  let $upButton = $(`#${id}`).children().children('.up-button');
-  let $downButton = $(`#${id}`).children().children('.down-button');
-  if (obj.completed === true) {
-    $completeButton.prop('disabled', true);
-    $downButton.prop('disabled', true);
-    $upButton.prop('disabled', true);
-  } else if (obj.importance === "5) None"){
-    $downButton.prop('disabled', true);
-    } else if (obj.importance === "4) Low" || obj.importance === "3) Normal" || obj.importance === "2) High") {
-      $downButton.prop('disabled', false);
-      $upButton.prop('disabled', false);
-      } else if (obj.importance === "1) Critical") {
-        $upButton.prop('disabled', true);
-      }
-}
+// function checkButtons(obj) {
+//   let id = obj.id;
+//   let $completeButton = $(`#${id}`).children().children('.completed-button');
+//   let $upButton = $(`#${id}`).children().children('.up-button');
+//   let $downButton = $(`#${id}`).children().children('.down-button');
+//   if (obj.completed === true) {
+//     $completeButton.prop('disabled', true);
+//     $downButton.prop('disabled', true);
+//     $upButton.prop('disabled', true);
+//   } else if (obj.importance === "5) None"){
+//     $downButton.prop('disabled', true);
+//     } else if (obj.importance === "4) Low" || obj.importance === "3) Normal" || obj.importance === "2) High") {
+//       $downButton.prop('disabled', false);
+//       $upButton.prop('disabled', false);
+//       } else if (obj.importance === "1) Critical") {
+//         $upButton.prop('disabled', true);
+//       }
+// }
+
 
 $('#search-box').keyup(function(){
    let filter = $(this).val(), count = 0;
@@ -270,11 +274,11 @@ $('#search-box').keyup(function(){
    });
  });
 
-function clearFields() {
-  $('#title-input').val('');
-  $('#task-input').val('');
-  $('#save-button').prop('disabled', true);
-}
+// function clearFields() {
+//   $('#title-input').val('');
+//   $('#task-input').val('');
+//   $('#save-button').prop('disabled', true);
+// }
 
 $('#title-input').on('input', function() {
   let length = 120 - $(this).val().length;
@@ -302,13 +306,13 @@ $('.all-input').keyup(function saveDisable() {
 $('.td-container').on('blur','.td-title', function() {
   let closestID = this.closest('article').id;
   let newTitle = $(this).text();
-  updateTDStorage(closestID, 'title', newTitle);
+  tdUpdate.updateTDStorage(closestID, 'title', newTitle);
 });
 
 $('.td-container').on('blur','.td-task', function() {
   let closestID = this.closest('article').id;
   let newBody = $(this).text();
-  updateTDStorage(closestID, 'task', newBody);
+  tdUpdate.updateTDStorage(closestID, 'task', newBody);
 });
 
 $('.td-container').on('keypress','.td-title', function(event) {
